requirements.txt
celery
coveralls
cryptography
django
django-celery-results
django-celery-beat
django-dbbackup
factory-boy
flower
gunicorn
pyasn1
pyopenssl
pyfakefs
pytest
pytest-django
pytest-cov
spur
feedparser
mock
python-coveralls
pymemcache
jsonrpc-requests
python-memcached
mysqlclient
ndg-httpsclient
raven
requests
responses
git+https://github.com/MiiRaGe/pyreport.git